+++ NaoTHSoccer/Source/Core/Cognition/Modules/Modeling/SelfLocator/MonteCarloSelfLocator/SampleSet.cpp
@@ -93,7 +93,7 @@ Sample SampleSet::meanOfLargestCluster(Moments2<2>& moments)
 
   for(unsigned int i = 0; i < numberOfParticles; i++)
   {
-    if(samples[i].cluster >= 0 && samples[i].cluster < numberOfParticles)
+    if(samples[i].cluster >= 0 && samples[i].cluster < (int)numberOfParticles)
     {
       int idx = samples[i].cluster;
       cluster[idx]++;
